[nixos] multiple improvements on UX for GNOME on Wayland
This adds a configuration to make sure the cursor is defined using Adwaita.

It configures the mouse focus to be sloppy, so when the mouse changes to
a different window, the window will be focused avoiding miss clicks.

It install Hyprland, Hyprland's compositor, and a few other packages to
start tinkering with it.

Installs and configures virtualization tools, including libvirt and
virtualbox to be used with vagrant.